+++ packages/core/src/types/general.ts
@@ -3,7 +3,7 @@ import type { MouseEvent as ReactMouseEvent, ComponentType, MemoExoticComponent 
 import type { D3DragEvent, Selection as D3Selection, SubjectPosition, ZoomBehavior } from 'd3';
 
 import type { XYPosition, Rect, Transform, CoordinateExtent } from './utils';
-import type { NodeChange, EdgeChange, NodePositionChange } from './changes';
+import type { NodeChange, EdgeChange } from './changes';
 import type {
   Node,
   NodeInternals,
@@ -155,6 +155,7 @@ export type ReactFlowStore = {
 
   nodesSelectionActive: boolean;
   userSelectionActive: boolean;
+  userSelectionRect: SelectionRect | null;
 
   connectionNodeId: string | null;
   connectionHandleId: string | null;
@@ -170,6 +171,7 @@ export type ReactFlowStore = {
   nodesFocusable: boolean;
   edgesFocusable: boolean;
   elementsSelectable: boolean;
+  elevateNodesOnSelect: boolean;
 
   multiSelectionActive: boolean;
 
@@ -212,6 +214,7 @@ export type ReactFlowStore = {
 
 export type ReactFlowActions = {
   setNodes: (nodes: Node[]) => void;
+  getNodes: () => Node[];
   setEdges: (edges: Edge[]) => void;
   setDefaultNodesAndEdges: (nodes?: Node[], edges?: Edge[]) => void;
   updateNodeDimensions: (updates: NodeDimensionUpdate[]) => void;
@@ -248,3 +251,13 @@ export type ProOptions = {
 };
 
 export type UseDragEvent = D3DragEvent<HTMLDivElement, null, SubjectPosition>;
+
+export enum SelectionMode {
+  Partial = 'partial',
+  Full = 'full',
+}
+
+export type SelectionRect = Rect & {
+  startX: number;
+  startY: number;
+};